added a simple test
